# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Enterprise Finance Dashboard API
  description: Multi-tenant B2B financial backend with granular RBAC and analytical aggregations.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users in the organization
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: search
          schema:
            type: string
          description: Fuzzy search by name or email
      responses:
        '200':
          description: List of users with pagination metadata
    post:
      tags:
        - Users
      summary: Invite a new employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name:
                  type: string
                  example: Peter Parker
                email:
                  type: string
                  example: peter@stark.com
                password:
                  type: string
                role:
                  type: string
                  enum: [ADMIN, ANALYST, VIEWER]
      responses:
        '201':
          description: Employee invited successfully
...
  /transactions:
    get:
      tags:
        - Transactions
      summary: Fetch financial ledger
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: type
          schema:
            type: string
            enum: [INCOME, EXPENSE]
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
            default: date
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of transactions with pagination
    post:
      tags:
        - Transactions
      summary: Create a financial record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, type, category]
              properties:
                amount:
                  type: number
                  example: 4500.50
                type:
                  type: string
                  enum: [INCOME, EXPENSE]
                category:
                  type: string
                  example: Software
                notes:
                  type: string
                  example: AWS Hosting
                date:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Transaction created
...